Form validation

+++ src/components/Unsubscribe.vue
@@ -2,39 +2,33 @@
   <div>
     <h1>Désinscription à la newsletter</h1>
     <div class="form-group">
-      <label for="email">Email à désinscrire</label>
-      <input type="email" class="form-control" id="email" placeholder="nom@domain.ch" v-model.lazy="email">
-      <button class="btn btn-primary" @click="unsubscribe" style="margin-top:10px;">Confirmer</button>
-      <div id="success-alert" class="alert alert-success" role="alert">
-        <i class="fa fa-check" aria-hidden="true"></i> Désinscrit!
-      </div>
-      <div id="error-alert" class="alert alert-danger" role="alert">
-        <i class="fa fa-times" aria-hidden="true"></i> Erreur
-      </div>
+      <custom-input ref="email" label="Email à désinscrire" type="text" v-model="email" errorMessage="L'email n'est pas valide." placeholder="nom@domain.ch" min="1" regexp="^\w+([\.-]?\w+)*@\w+([\.-]?\w+)*(\.\w{2,3})+$"></custom-input>
+      <custom-button ref="unsubscribeButton" @click="unsubscribe" text="Désinscrire" pendingText="Désinscription" successText="Désinscrit"></custom-button>
     </div>
   </div>
 </template>
 
 <script>
-import alert from '../alert'
 import api from '../../config/api.js'
 
 export default {
   data () {
     return {
-      email: null
+      email: ''
     }
-  },
-  mounted: function () {
-    alert.hideAll()
   },
   methods: {
     unsubscribe () {
+      const isEmailValid = this.$refs.email.isValid(this.email)
+      if (!isEmailValid) {
+        this.$refs.unsubscribeButton.showError(`L'email n'est pas valide.`)
+        return
+      }
       this.$http.delete(`${api.subscriptions}/${this.email}`)
       .then((response) => {
-        alert.show('#success-alert')
+        this.$refs.unsubscribeButton.showSuccess()
       }, (response) => {
-        alert.show('#error-alert')
+        this.$refs.unsubscribeButton.showError()
       })
     }
   }
